Add note about how to create resource based policy
As a relative noob, I wrestled with AWS IAM policies for some time before I discovered the note in the documentation stating that resource based policies can only be created for SM through the CLI or using API's. I added a note linking to that documentation where I would have seen it quickly.